# service/openapi.yaml
openapi: 3.0.3
info:
  title: Respire sync server
  version: 1.0.0
  description: >
    Ciphertext store plus auth. The server never decrypts payloads.
    CLI and other clients talk to these paths. This is a partial sync/self API
    schema; Admin JSON routes are not yet fully described. Frontends are owned by
    respire-site. See ../docs/browser-api-contract.md for the cross-origin
    GET/POST + Authorization/Content-Type contract and staged rollout.
servers:
  - url: https://api.rsrs.rs
    description: Assigned Respire API endpoint; configuration only, not proof of DNS or deployment.
  - url: http://127.0.0.1:8787
    description: Local development server.
paths:
...
  /login:
    post:
      summary: Password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user, pass_hash]
              properties:
                user: { type: string }
                pass_hash: { type: string }
                device_name: { type: string }
      responses:
        "200": { description: Session issued }
        "401": { description: Bad credentials }
...
  /forget:
    post:
      summary: Tombstone one id
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: deleted true/false }
  /count:
    get:
      summary: Alive blob count
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Count }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer